requirements.txt
python-dotenv==1.0.0
python-telegram-bot==13.0
requests==2.31.0


